requirements.txt
streamlit
praw
pandas
openpyxl
requests
markdown2
pdfkit
python-dotenv
